requirements.txt
tensorflow>=2.15.0
keras>=3.0.0
numpy>=1.23.5
matplotlib>=3.7.1
scikit-learn>=1.2.2
scipy>=1.10.1
Pillow>=9.4.0
pandas>=2.0.0
opencv-python   
streamlit>=1.20.0
